pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mystquarto"
version = "0.1.0"
description = "Bidirectional MyST ↔ Quarto converter"
readme = "README.md"
license = "MIT-0"
requires-python = ">=3.10"
authors = [{ name = "Max Ghenis" }]
keywords = ["myst", "quarto", "markdown", "converter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT No Attribution License (MIT-0)",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.4"]

[project.scripts]
myst2quarto = "mystquarto.cli:myst2quarto"
quarto2myst = "mystquarto.cli:quarto2myst"
mystquarto = "mystquarto.cli:main"

[project.urls]
Homepage = "https://github.com/MaxGhenis/mystquarto"
Repository = "https://github.com/MaxGhenis/mystquarto"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
